refactor openai and claude oversized modules
