+++ packages/claude-agent-sdk-haskell/src/Claude/Agent/SDK/Client.hs
@@ -27,6 +27,11 @@ import Claude.Agent.SDK.Errors
 import Claude.Agent.SDK.Internal.MessageParser
     ( decodeMessageLine
     )
+import Claude.Agent.SDK.Internal.Client.Usage
+    ( UsageAccounting(..)
+    , cumulativeUsage
+    , reconcileCumulativeUsage
+    )
 import Claude.Agent.SDK.Internal.Transport.SubprocessCLI
     ( newSubprocessCLITransport
     )
@@ -44,7 +49,6 @@ import Claude.Agent.SDK.Types
     , Usage(..)
     , addUsage
     , emptyUsage
-    , modelUsageToUsage
     )
 import Control.Applicative ((<|>))
 import Control.Concurrent (threadDelay)
@@ -111,17 +115,6 @@ data RunningClient = RunningClient
     , runningUsageAccounting :: !(IORef UsageAccounting)
     , runningClosed :: !(IORef Bool)
     , runningCloseLock :: !(MVar ())
-    }
-
-data UsageAccounting = UsageAccounting
-    { usageCumulativeBaseline :: !(Maybe Usage)
-    , usagePendingFallback :: !Usage
-    }
-
-data UsageComponents = UsageComponents
-    { usageUncachedInput :: !Int
-    , usageCachedInput :: !Int
-    , usageOutput :: !Int
     }
 
 data ClaudeSDKTurn = ClaudeSDKTurn
@@ -932,112 +925,6 @@ connectionError prefix exception =
         Nothing ->
             CLIConnectionError
                 (prefix <> ": " <> Text.pack (show exception))
-
-cumulativeUsage :: Map.Map Text ModelUsage -> Maybe Usage
-cumulativeUsage modelUsage =
-    case Map.elems modelUsage of
-        [] -> Nothing
-        entries ->
-            Just $
-                foldl
-                    addUsage
-                    emptyUsage
-                    (map modelUsageToUsage entries)
-
-usageDelta :: Usage -> Usage -> Usage
-usageDelta previous current
-    | usageIsMonotonic previous current =
-        fromUsageComponents $
-            subtractUsageComponents
-                (toUsageComponents previous)
-                (toUsageComponents current)
-    | otherwise =
-        current
-
-reconcileCumulativeUsage
-    :: UsageAccounting
-    -> Usage
-    -> (Usage, Usage)
-reconcileCumulativeUsage accounting current =
-    case accounting.usageCumulativeBaseline of
-        Just previous
-            | not (usageIsMonotonic previous current) ->
-                (current, emptyUsage)
-        previous ->
-            subtractReportedFallback
-                accounting.usagePendingFallback
-                (maybe current (`usageDelta` current) previous)
-
-usageIsMonotonic :: Usage -> Usage -> Bool
-usageIsMonotonic previous current =
-    let previousComponents = toUsageComponents previous
-        currentComponents = toUsageComponents current
-    in
-        currentComponents.usageUncachedInput
-            >= previousComponents.usageUncachedInput
-            && currentComponents.usageCachedInput
-                >= previousComponents.usageCachedInput
-            && currentComponents.usageOutput
-                >= previousComponents.usageOutput
-
-subtractReportedFallback
-    :: Usage
-    -> Usage
-    -> (Usage, Usage)
-subtractReportedFallback pending gross =
-    ( fromUsageComponents $
-        subtractUsageComponents
-            (toUsageComponents pending)
-            (toUsageComponents gross)
-    , fromUsageComponents $
-        subtractUsageComponents
-            (toUsageComponents gross)
-            (toUsageComponents pending)
-    )
-
-toUsageComponents :: Usage -> UsageComponents
-toUsageComponents usage =
-    let cached =
-            max 0 (min usage.inputTokens usage.cachedTokens)
-    in UsageComponents
-        { usageUncachedInput =
-            max 0 (usage.inputTokens - cached)
-        , usageCachedInput = cached
-        , usageOutput = max 0 usage.outputTokens
-        }
-
-fromUsageComponents :: UsageComponents -> Usage
-fromUsageComponents components =
-    Usage
-        { inputTokens =
-            components.usageUncachedInput
-                + components.usageCachedInput
-        , outputTokens = components.usageOutput
-        , cachedTokens = components.usageCachedInput
-        }
-
-subtractUsageComponents
-    :: UsageComponents
-    -> UsageComponents
-    -> UsageComponents
-subtractUsageComponents subtrahend minuend =
-    UsageComponents
-        { usageUncachedInput =
-            max 0
-                ( minuend.usageUncachedInput
-                    - subtrahend.usageUncachedInput
-                )
-        , usageCachedInput =
-            max 0
-                ( minuend.usageCachedInput
-                    - subtrahend.usageCachedInput
-                )
-        , usageOutput =
-            max 0
-                ( minuend.usageOutput
-                    - subtrahend.usageOutput
-                )
-        }
 
 nonEmptyText :: Maybe Text -> Maybe Text
 nonEmptyText =
